Fix closing of dialogs triggered by hd_input
After the upgrade to jquery-ui 1.10, closing a dialog bound to a
`hd_input` didn't work properly: now the closing of the dialog
apparently triggers another focus event for the display-field of the
`hd_input`.

Change the dialogs used for `hd_input` fields to avoid this behavior:

* Bind `beforeClose` of the dialog to set a flag that's considered by
  `hd_input_focus`

* Bind `close` of the dialog to clear that flag

* Use `.dialog ("close")` to close the dialog programmatically

This way, it doesn't matter how the dialog is closed (by the `Cancel`
button, by `Escape`, or by the `Close` button supplied by jquery-ui
itself). In all cases, the flag is set/cleared and `hd_input_focus`
doesn't try to re-open the dialog during closing.