Use `.attr_prop`  to access attribute descriptors
Up to now, most accesses to attribute descriptors were done applying
``getattr`` to the E_Type. This makes it impossible to define attribute
descriptors that return a value when applied to the class.

By using ``.attr_prop`` to access attribute descriptors, it is possible
to define attribute kinds for attributes like `type_name` or
``x_locked`` that provide access to the attribute value via the class: a
constant for ``type_name``, the default value (possibly constant) for
``x_locked``.