Do not use sequence tables in SAW for pid, cid
SAW.Sequence used to use a separate table for each sequence to allow
recovery of the max value even after rollback and delete operations.
This causes a massive performance hit for PostgreSQL.

Now, the sequences for `pid` and `cid` don't use a separate table for
the sequence. `max_pid` and `max_cid` are now implemented as queries
over the table for `MOM.Id_Entity` and `MOM.MD_Change`. This increases
performance for the price of backend-specific values after rollbacks
(which mainly hurts in the tests).