Use `python -m` to run `_TFL.Babel`
If `babel.sh compile` is called for an application with the same name as
a module in `_TFL`, the import of the application will not be done if
`_TFL` is in `sys.path`.

`python -m _TFL.Babel` won't add `.../_TFL` to `sys.path` whereas
using `python ${lib}/_TFL/Babel.py` does.