Add `show_in_ui_selector` to MOM.Attr.Type
Setting `show_in_ui_selector` to False hides that attribute from
UI selector dialogs for types containing that attribute.

Beware: that might make entity selection difficult if the hidden
attribute is necessary to restrict the number of matching entities to a
displayable set (it's not possible to show more than a few tens of
entities for selection).