Add guard `self._request.user` to `username`
For anonymous edit sessions, e.g. edit sessions for a user that hasn't
logged in, a session with a random, but unique, `username` is created to
store the edit session.

As that `username` doesn't match any account, it should **not** be
displayed in a login form; the newly added guard ensures that.